+++ src/pytest_cases/fixture_core1_unions.py
@@ -239,14 +239,14 @@ def ignore_unused(fixture_func):
         new_sig = old_sig
 
     if isasyncgenfunction(fixture_func) and sys.version_info >= (3.6):
-        from .fixture_core_pep525 import _ignore_unused_asyncgen_pep525
+        from .pep525 import _ignore_unused_asyncgen_pep525
         wrapped_fixture_func = _ignore_unused_asyncgen_pep525(fixture_func, new_sig, func_needs_request)
     elif iscoroutinefunction(fixture_func) and sys.version_info >= (3.5):
-        from .fixture_core_pep492 import _ignore_unused_coroutine_pep492
+        from .pep492 import _ignore_unused_coroutine_pep492
         wrapped_fixture_func = _ignore_unused_coroutine_pep492(fixture_func, new_sig, func_needs_request)
     elif isgeneratorfunction(fixture_func):
         if sys.version_info >= (3, 3):
-            from .fixture_core_pep380 import _ignore_unused_generator_pep380
+            from .pep380 import _ignore_unused_generator_pep380
             wrapped_fixture_func = _ignore_unused_generator_pep380(fixture_func, new_sig, func_needs_request)
         else:
         # generator function (with a yield statement)
